redzone: add support for continuous compilation.
this change introduces support for automated zone compilation
whenever the configuration changes. it can make use of inotify
events (if supported) or simply poll the file for changes.

BREAKING CHANGE:
the `generate-zonefiles` script has been replaced with an action
argument of the same name to the docker-entrypoint. intervention
is only necessary if you called the script with an absolute path
before.