powerdns-admin: extend *PYTHONPATH* with config directory.
the app configuration can be placed freely in the filesystem using
the *POWERDNS_ADMIN_CONF* environment variable. the default path
is inside the app source tree, which allows the scripts to **import**
the file without any changes to the environment. should the path
change, the app needs to be informed about the new location. adding
the directory to the *PYTHONPATH* allows all script and the main
application to find the file easily.

fixes #5